create a comment component
